test: move vim.iter tests to separate file
